`flatMapIt()` no longer calls `flatMapArray()`
