+++ src/transformation/flat-map-it.ts
@@ -2,17 +2,12 @@
  * @packageDocumentation
  * @module @proc7ts/push-iterator
  */
-import { overIterable, overNone } from '../construction';
 import { itsIterator } from '../its-iterator';
 import { makePushIterator } from '../make-push-iterator';
 import type { PushIterable, PushOrRawIterable } from '../push-iterable';
 import { isPushIterable, PushIterable__symbol } from '../push-iterable';
 import type { PushIterator } from '../push-iterator';
-
-/**
- * @internal
- */
-const flatMapIt$defaultConverter = <T, R>(element: T): PushOrRawIterable<R> => element as unknown as Iterable<R>;
+import { flatMapIt$defaultConverter } from './transformation.impl';
 
 /**
  * Flattens the source iterable of iterables into new {@link PushIterable push iterable}.
@@ -46,13 +41,7 @@ export function flatMapIt<T, R>(
     source: PushOrRawIterable<T>,
     convert: (this: void, element: T) => PushOrRawIterable<R> = flatMapIt$defaultConverter,
 ): PushIterable<R> {
-  if (isPushIterable(source)) {
-    return flatMapPushIterable(source, convert);
-  }
-  if (Array.isArray(source)) {
-    return flatMapArray(source, convert);
-  }
-  return flatMapRawIterable(source, convert);
+  return isPushIterable(source) ? flatMapPushIterable(source, convert) : flatMapRawIterable(source, convert);
 }
 
 /**
@@ -92,73 +81,6 @@ function flatMapPushIterable<T, R>(
             if (lastSrc) {
               return false;
             }
-          }
-          if (goOn === false) {
-            return true;
-          }
-        }
-      });
-    },
-  };
-}
-
-/**
- * Flattens the source `array` of iterables into new {@link PushIterable push iterable}.
- *
- * Calling this function is the same as calling `flatMapArray(source, element => element)`.
- *
- * @typeParam T  A type of converted elements.
- * @param array  A source array-like instance of iterables.
- *
- * @returns New push iterable with each element of `array` being the flattened.
- */
-export function flatMapArray<T>(array: ArrayLike<Iterable<T>>): PushIterable<T>;
-
-/**
- * First maps each element of the source `array` using a mapping function, then flattens the result into new
- * {@link PushIterable push iterable}.
- *
- * @typeParam T  A type of array elements.
- * @typeParam R  A type of converted elements.
- * @param array  A source array-like instance of iterables.
- * @param convert  A function that produces new iterable, taking array element as the only parameter.
- *
- * @returns New push iterable with each element being the flattened result of the `convert` function call.
- */
-export function flatMapArray<T, R>(
-    array: ArrayLike<T>,
-    convert: (this: void, element: T) => PushOrRawIterable<R>,
-): PushIterable<R>;
-
-export function flatMapArray<T, R>(
-    array: ArrayLike<T>,
-    convert: (this: void, element: T) => PushOrRawIterable<R> = flatMapIt$defaultConverter,
-): PushIterable<R> {
-  if (array.length <= 1) {
-    if (!array.length) {
-      return overNone();
-    }
-    return overIterable(convert(array[0]));
-  }
-
-  return {
-    [PushIterable__symbol]: 1,
-    [Symbol.iterator]() {
-
-      let cIt: PushIterator<R> = itsIterator(convert(array[0]));
-      let index = 1;
-
-      return makePushIterator(accept => {
-        for (; ;) {
-
-          // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
-          let goOn: boolean | void;
-
-          if (!cIt.forNext(element => goOn = accept(element))) {
-            if (index >= array.length) {
-              return false;
-            }
-            cIt = itsIterator(convert(array[index++]));
           }
           if (goOn === false) {
             return true;
