+++ src/transformation/index.ts
@@ -1,5 +1,6 @@
 export * from './filter-array';
 export * from './filter-it';
+export * from './flat-map-array';
 export * from './flat-map-it';
 export * from './map-array';
 export * from './map-it';
