+++ src/transformation/transformation.impl.ts
@@ -1,0 +1,6 @@
+import type { PushOrRawIterable } from '../push-iterable';
+
+/**
+ * @internal
+ */
+export const flatMapIt$defaultConverter = <T, R>(element: T): PushOrRawIterable<R> => element as unknown as Iterable<R>;
